refactor: Move RiskEngine out of models/ and remove runner-level cooldown duplication
