+++ tests/Nightwatch/test_strategy_runner.py
@@ -2,13 +2,12 @@
 """Unit tests for the StrategyRunner class."""
 
 import unittest
-from collections import deque
 from datetime import datetime, timedelta, timezone
 from decimal import Decimal
 
 from Nightwatch.metrics import NightwatchMetrics
-from Nightwatch.models.risk_engine import RiskEngine
 from Nightwatch.models.tick_buffer import TickBuffer
+from Nightwatch.risk_engine import RiskEngine
 from Nightwatch.rules.cooldown_rule import CooldownRule
 from Nightwatch.rules.max_signal_per_minute_rule import MaxSignalPerMinuteRule
 from Nightwatch.rules.min_trade_strength_rule import MinTradeStrengthRule
@@ -28,7 +27,6 @@ class TestStrategyRunner(unittest.TestCase):
         self.runner = StrategyRunner(
             strategy=self.__strategy,
             buffer=self.__buffer,
-            cooldown=timedelta(seconds=10),
             metric=self._metric,
             risk_engine=self.risk_engine,
         )
@@ -50,59 +48,6 @@ class TestStrategyRunner(unittest.TestCase):
         )
         signal = feed_ticks(self.runner, failing_ticks)
         self.assertIsNone(signal)
-
-    def test_cooldown_prevents_signal(self) -> None:
-        """Test that the cooldown period prevents emitting signals too frequently."""
-        strategy_no_cd = MomentumBurstStrategy(threshold_pct=1, window_sec=60, metric=self._metric)
-        buffer_no_cd = TickBuffer(max_ticks_per_symbol=30)
-        runner_no_cd = StrategyRunner(strategy=strategy_no_cd, buffer=buffer_no_cd, risk_engine=self.risk_engine)
-
-        ticks = make_tick_sequence(
-            prices=[Decimal("100"), Decimal("105"), Decimal("115")], start=self.start_time, interval_sec=5.0, symbol="BTC/USD"
-        )
-
-        signal = feed_ticks(runner_no_cd, ticks)
-        self.assertIsNotNone(signal)
-        self.assertEqual(signal.side.value, "BUY")  # type: ignore[union-attr]
-
-        next_tick = make_tick(price=Decimal("130"), timestamp=self.start_time + timedelta(seconds=11))
-        second_signal_no_cd = runner_no_cd.on_market_tick(next_tick)
-        self.assertIsNotNone(second_signal_no_cd)
-        self.assertEqual(second_signal_no_cd.side.value, "BUY")
-
-        strategy_cd = MomentumBurstStrategy(threshold_pct=1, window_sec=60, metric=self._metric)
-        buffer_cd = TickBuffer(max_ticks_per_symbol=30)
-        runner_cd = StrategyRunner(strategy=strategy_cd, buffer=buffer_cd, cooldown=timedelta(seconds=10), risk_engine=self.risk_engine)
-
-        signal = None
-        i = 0
-        while signal is None and i < len(ticks):
-            tick = ticks[i]
-            signal = runner_cd.on_market_tick(tick)
-            i += 1
-        self.assertIsNotNone(signal)
-        self.assertEqual(signal.side.value, "BUY")  # type: ignore[union-attr]
-
-        signal_during_cooldown = runner_cd.on_market_tick(next_tick)
-        self.assertIsNone(signal_during_cooldown)
-
-    def test_cooldown_increments_suppressed_metric(self) -> None:
-        """Given a signal was just emitted, when the next tick arrives within cooldown,
-        then signals_suppressed_total{reason='cooldown'} increments by 1."""
-        ticks = make_tick_sequence(
-            prices=[Decimal("100"), Decimal("105"), Decimal("115")], start=self.start_time, interval_sec=5.0, symbol="BTC/USD"
-        )
-        for tick in ticks:
-            self.runner.on_market_tick(tick)
-        suppressed_tick = make_tick(
-            price=Decimal("130"),
-            timestamp=self.start_time + timedelta(seconds=15),
-        )
-        result = self.runner.on_market_tick(suppressed_tick)
-
-        self.assertIsNone(result)
-        value = self._metric.get_counter_value(self._metric.signals_suppressed_total, reason="cooldown")
-        self.assertEqual(value, 1.0)
 
     def test_integration_with_momentum_burst_strategy(self) -> None:
         """Test that the StrategyRunner correctly integrates with the MomentumBurstStrategy."""
@@ -133,39 +78,6 @@ class TestStrategyRunner(unittest.TestCase):
         self.assertIn("window_sec", log_output)
         self.assertIn("threshold_pct", log_output)
 
-    def test_cooldown_per_symbol(self) -> None:
-        """Test that the cooldown is applied separately for each symbol."""
-        metric = NightwatchMetrics()
-        strategy = MomentumBurstStrategy(threshold_pct=10, metric=metric)
-        buffer = TickBuffer(max_ticks_per_symbol=30)
-        runner = StrategyRunner(strategy=strategy, buffer=buffer, cooldown=timedelta(seconds=30), metric=metric)
-        ticks_symbol = deque(
-            [
-                make_tick(price=Decimal("100"), timestamp=self.start_time, symbol="A"),
-                make_tick(price=Decimal("100"), timestamp=self.start_time, symbol="B"),
-                make_tick(price=Decimal("105"), timestamp=self.start_time + timedelta(seconds=5), symbol="A"),
-                make_tick(price=Decimal("105"), timestamp=self.start_time + timedelta(seconds=5), symbol="B"),
-                make_tick(price=Decimal("115"), timestamp=self.start_time + timedelta(seconds=10), symbol="A"),
-                make_tick(price=Decimal("115"), timestamp=self.start_time + timedelta(seconds=10), symbol="B"),
-                make_tick(price=Decimal("130"), timestamp=self.start_time + timedelta(seconds=15), symbol="A"),
-                make_tick(price=Decimal("130"), timestamp=self.start_time + timedelta(seconds=45), symbol="B"),
-            ]
-        )
-        signal_a = None
-        signal_b = None
-        for tick in ticks_symbol:
-            result = runner.on_market_tick(tick)
-            if tick.symbol == "A" and signal_a is None and result is not None:
-                signal_a = result
-            elif tick.symbol == "B" and signal_b is None and result is not None:
-                signal_b = result
-
-        self.assertIsNotNone(signal_a)
-        self.assertIsNotNone(signal_b)
-        self.assertEqual(signal_a.side, "BUY")  # type: ignore[union-attr]
-        self.assertEqual(signal_b.side, "BUY")  # type: ignore[union-attr]
-        self.assertEqual(metric.get_counter_value(metric.signals_suppressed_total, reason="cooldown"), 1.0)
-
     def test_signals_total(self) -> None:
         """Test that the counter signals_total is incremented correctly."""
         initial_evaluations = (
@@ -185,9 +97,7 @@ class TestStrategyRunner(unittest.TestCase):
         then both emit signals (no suppression)."""
         strategy = MomentumBurstStrategy(threshold_pct=10, metric=self._metric)
         buffer = TickBuffer(max_ticks_per_symbol=30)
-        runner = StrategyRunner(
-            strategy=strategy, buffer=buffer, cooldown=timedelta(seconds=0), metric=self._metric, risk_engine=self.risk_engine
-        )
+        runner = StrategyRunner(strategy=strategy, buffer=buffer, metric=self._metric, risk_engine=self.risk_engine)
         ticks = make_tick_sequence(prices=[Decimal("100"), Decimal("115")], start=self.start_time, interval_sec=10.0, symbol="BTC/USD")
         signal1 = feed_ticks(runner, ticks)
         self.assertIsNotNone(signal1)
@@ -203,7 +113,7 @@ class TestStrategyRunner(unittest.TestCase):
         strategy = MomentumBurstStrategy(threshold_pct=10, metric=metric)
         buffer = TickBuffer(max_ticks_per_symbol=30)
         risk_engine = RiskEngine(rules=[MinTradeStrengthRule(min_strength=50.0)])
-        runner = StrategyRunner(strategy=strategy, buffer=buffer, cooldown=timedelta(seconds=0), metric=metric, risk_engine=risk_engine)
+        runner = StrategyRunner(strategy=strategy, buffer=buffer, metric=metric, risk_engine=risk_engine)
 
         # delta_pct = 15%, strength = 15.0, which is below min_strength=50
         ticks = make_tick_sequence(
@@ -218,7 +128,7 @@ class TestStrategyRunner(unittest.TestCase):
         strategy = MomentumBurstStrategy(threshold_pct=10, metric=metric)
         buffer = TickBuffer(max_ticks_per_symbol=30)
         risk_engine = RiskEngine(rules=[MinTradeStrengthRule(min_strength=50.0)])
-        runner = StrategyRunner(strategy=strategy, buffer=buffer, cooldown=timedelta(seconds=0), metric=metric, risk_engine=risk_engine)
+        runner = StrategyRunner(strategy=strategy, buffer=buffer, metric=metric, risk_engine=risk_engine)
 
         ticks = make_tick_sequence(
             prices=[Decimal("100"), Decimal("105"), Decimal("115")], start=self.start_time, interval_sec=5.0, symbol="BTC/USD"
@@ -236,7 +146,7 @@ class TestStrategyRunner(unittest.TestCase):
         strategy = MomentumBurstStrategy(threshold_pct=10, metric=metric)
         buffer = TickBuffer(max_ticks_per_symbol=30)
         risk_engine = RiskEngine(rules=[MinTradeStrengthRule(min_strength=50.0)])
-        runner = StrategyRunner(strategy=strategy, buffer=buffer, cooldown=timedelta(seconds=0), metric=metric, risk_engine=risk_engine)
+        runner = StrategyRunner(strategy=strategy, buffer=buffer, metric=metric, risk_engine=risk_engine)
 
         with self.assertLogs("Nightwatch.strategy_runner", level="INFO") as log:
             ticks = make_tick_sequence(
@@ -256,7 +166,7 @@ class TestStrategyRunner(unittest.TestCase):
         strategy = MomentumBurstStrategy(threshold_pct=10, window_sec=60, metric=metric)
         buffer = TickBuffer(max_ticks_per_symbol=30)
         risk_engine = RiskEngine(rules=[MaxSignalPerMinuteRule(max_signals_per_min=1)])
-        runner = StrategyRunner(strategy=strategy, buffer=buffer, cooldown=timedelta(seconds=0), metric=metric, risk_engine=risk_engine)
+        runner = StrategyRunner(strategy=strategy, buffer=buffer, metric=metric, risk_engine=risk_engine)
 
         ticks = make_tick_sequence(prices=[Decimal("100"), Decimal("115")], start=self.start_time, interval_sec=5.0, symbol="BTC/USD")
         signal1 = feed_ticks(runner, ticks)
@@ -276,7 +186,7 @@ class TestStrategyRunner(unittest.TestCase):
         strategy = MomentumBurstStrategy(threshold_pct=10, metric=metric)
         buffer = TickBuffer(max_ticks_per_symbol=30)
         risk_engine = RiskEngine(rules=[MinTradeStrengthRule(min_strength=1.0), MaxSignalPerMinuteRule(max_signals_per_min=1000)])
-        runner = StrategyRunner(strategy=strategy, buffer=buffer, cooldown=timedelta(seconds=0), metric=metric, risk_engine=risk_engine)
+        runner = StrategyRunner(strategy=strategy, buffer=buffer, metric=metric, risk_engine=risk_engine)
 
         # delta_pct = 15%, strength = 15.0 — above min_strength=1.0, within rate limit
         ticks = make_tick_sequence(
@@ -295,7 +205,7 @@ class TestStrategyRunner(unittest.TestCase):
         strategy = MomentumBurstStrategy(threshold_pct=1, window_sec=60, metric=metric)
         buffer = TickBuffer(max_ticks_per_symbol=30)
         # No risk_engine provided — default includes MinTradeStrengthRule(min_strength=10)
-        runner = StrategyRunner(strategy=strategy, buffer=buffer, cooldown=timedelta(seconds=0), metric=metric)
+        runner = StrategyRunner(strategy=strategy, buffer=buffer, metric=metric)
 
         # delta_pct = 5%, strength = 5.0, below default min_strength=10
         ticks = make_tick_sequence(prices=[Decimal("100"), Decimal("105")], start=self.start_time, interval_sec=5.0, symbol="BTC/USD")
@@ -312,7 +222,7 @@ class TestStrategyRunner(unittest.TestCase):
         buffer = TickBuffer(max_ticks_per_symbol=30)
         # CooldownRule is first; MinTradeStrengthRule second
         risk_engine = RiskEngine(rules=[CooldownRule(cooldown_seconds=999), MinTradeStrengthRule(min_strength=50.0)])
-        runner = StrategyRunner(strategy=strategy, buffer=buffer, cooldown=timedelta(seconds=0), metric=metric, risk_engine=risk_engine)
+        runner = StrategyRunner(strategy=strategy, buffer=buffer, metric=metric, risk_engine=risk_engine)
 
         # Signal 1: delta_pct=55% → strength=55 ≥ 50 → passes both rules → CooldownRule confirm() called
         ticks = make_tick_sequence(prices=[Decimal("100"), Decimal("155")], start=self.start_time, interval_sec=5.0, symbol="BTC/USD")
